base code for video rental demo
