+++ src/main/java/video/rental/demo/Main.java
@@ -1,11 +1,17 @@
 package video.rental.demo;
 
+import video.rental.demo.application.Interactor;
+import video.rental.demo.domain.Repository;
+import video.rental.demo.infrastructure.RepositoryMemImpl;
+import video.rental.demo.presentation.CmdUI;
+
 public class Main {
 	private static CmdUI ui;
 
 	public static void main(String[] args) {
-//		Repository repository = new RepositoryMemImpl();
-//		ui = new CmdUI(repository);
-//		ui.start();
+		Repository repository = new RepositoryMemImpl();
+		Interactor interactor = new Interactor(repository);
+		ui = new CmdUI(interactor);
+		ui.start();
 	}
 }
