+++ src/main/java/video/rental/demo/application/Interactor.java
@@ -1,0 +1,158 @@
+package video.rental.demo.application;
+
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.List;
+
+import video.rental.demo.domain.Customer;
+import video.rental.demo.domain.Rating;
+import video.rental.demo.domain.Rental;
+import video.rental.demo.domain.Repository;
+import video.rental.demo.domain.Video;
+
+public class Interactor {
+
+	private Repository repository;
+
+	public Interactor(Repository repository) {
+		super();
+		this.repository = repository;
+	}
+
+	public void clearRentals(int customerCode) {
+		Customer foundCustomer = getRepository().findCustomerById(customerCode);
+	
+		if (foundCustomer == null) {
+			System.out.println("No customer found");
+		} else {
+			System.out.println("Id: " + foundCustomer.getCode() + "\nName: " + foundCustomer.getName() + "\tRentals: "
+					+ foundCustomer.getRentals().size());
+			for (Rental rental : foundCustomer.getRentals()) {
+				System.out.print("\tTitle: " + rental.getVideo().getTitle() + " ");
+				System.out.print("\tPrice Code: " + rental.getVideo().getPriceCode());
+			}
+	
+			List<Rental> rentals = new ArrayList<Rental>();
+			foundCustomer.setRentals(rentals);
+	
+			getRepository().saveCustomer(foundCustomer);
+		}
+	}
+
+	public void returnVideo(int customerCode, String videoTitle) {
+		Customer foundCustomer = getRepository().findCustomerById(customerCode);
+		if (foundCustomer == null)
+			return;
+	
+		List<Rental> customerRentals = foundCustomer.getRentals();
+	
+		for (Rental rental : customerRentals) {
+			if (rental.getVideo().getTitle().equals(videoTitle) && rental.getVideo().isRented()) {
+				Video video = rental.returnVideo();
+				video.setRented(false);
+				getRepository().saveVideo(video);
+				break;
+			}
+		}
+	
+		getRepository().saveCustomer(foundCustomer);
+	}
+
+	public void listVideos() {
+		List<Video> videos = getRepository().findAllVideos();
+	
+		for (Video video : videos) {
+			System.out.println(
+					"Video type: " + video.getVideoType() + 
+					"\tPrice code: " + video.getPriceCode() + 
+					"\tRating: " + video.getVideoRating() +
+					"\tTitle: " + video.getTitle()
+					); 
+		}
+	}
+
+	public void listCustomers() {
+		List<Customer> customers = getRepository().findAllCustomers();
+	
+		for (Customer customer : customers) {
+			System.out.println("ID: " + customer.getCode() + "\nName: " + customer.getName() + "\tRentals: "
+					+ customer.getRentals().size());
+			for (Rental rental : customer.getRentals()) {
+				System.out.print("\tTitle: " + rental.getVideo().getTitle() + " ");
+				System.out.print("\tPrice Code: " + rental.getVideo().getPriceCode());
+				System.out.println("\tReturn Status: " + rental.getStatus());
+			}
+		}
+	}
+
+	public void getCustomerReport(int code) {
+		Customer foundCustomer = getRepository().findCustomerById(code);
+	
+		if (foundCustomer == null) {
+			System.out.println("No customer found");
+		} else {
+			String result = foundCustomer.getReport();
+			System.out.println(result);
+		}
+	}
+
+	public void rentVideo(int code, String videoTitle) {
+		Customer foundCustomer = getRepository().findCustomerById(code);
+		if (foundCustomer == null)
+			return;
+	
+		Video foundVideo = getRepository().findVideoByTitle(videoTitle);
+	
+		if (foundVideo == null)
+			return;
+	
+		if (foundVideo.isRented() == true)
+			return;
+	
+		Boolean status = foundVideo.rentFor(foundCustomer);
+		if (status == true) {
+			getRepository().saveVideo(foundVideo);
+			getRepository().saveCustomer(foundCustomer);
+		} else {
+			return;
+		}
+	}
+
+	public void registerCustomer(String name, int code, String dateOfBirth) {
+		Customer customer = new Customer(code, name, LocalDate.parse(dateOfBirth));
+		getRepository().saveCustomer(customer);
+	}
+	
+	// User가 register하는 것이 아닌 괌점에서는 Date는 Presentation 로직이 아니다.
+	public void registerVideo(String title, int videoType, int priceCode, int videoRating, LocalDate registeredDate) {
+		Rating rating;
+		if (videoRating == 1) rating = Rating.TWELVE;
+		else if (videoRating == 2) rating = Rating.FIFTEEN;
+		else if (videoRating == 3) rating = Rating.EIGHTEEN;
+		else throw new IllegalArgumentException("No such rating " + videoRating);
+		
+		Video video = new Video(title, videoType, priceCode, rating, registeredDate);
+	
+		getRepository().saveVideo(video);
+	}
+
+	// Testability 관점에서 내부적으로 시간을 설정하는 것은 불편하다.
+	public void registerVideo(String title, int videoType, int priceCode, int videoRating) {
+		LocalDate registeredDate = LocalDate.now();
+		
+		Rating rating;
+		if (videoRating == 1) rating = Rating.TWELVE;
+		else if (videoRating == 2) rating = Rating.FIFTEEN;
+		else if (videoRating == 3) rating = Rating.EIGHTEEN;
+		else throw new IllegalArgumentException("No such rating " + videoRating);
+		
+		Video video = new Video(title, videoType, priceCode, rating, registeredDate);
+	
+		getRepository().saveVideo(video);
+	}
+
+	Repository getRepository() {
+		return repository;
+	}
+
+}
