+++ src/main/java/video/rental/demo/domain/Repository.java
@@ -1,4 +1,4 @@
-package video.rental.demo;
+package video.rental.demo.domain;
 
 import java.util.List;
