+++ src/main/java/video/rental/demo/presentation/GraphicUI.java
@@ -1,3 +1,4 @@
+package video.rental.demo.presentation;
 //package video.rental.demo;
 //
 //import java.awt.EventQueue;
