+++ src/test/java/video/rental/demo/GoldenMaster.java
@@ -8,6 +8,12 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.stream.Collectors;
 
+import video.rental.demo.application.Interactor;
+import video.rental.demo.domain.Repository;
+import video.rental.demo.infrastructure.RepositoryMemImpl;
+import video.rental.demo.presentation.CmdUI;
+import video.rental.demo.util.SampleGenerator;
+
 public class GoldenMaster {
 	
 	private static final String GoldenMasterFile = "./goldenmaster/goldenmaster.txt";
@@ -41,8 +47,9 @@ public class GoldenMaster {
 		ByteArrayOutputStream ostream = redirectOutput();
 		
 		Repository repository = new RepositoryMemImpl();
+		Interactor interactor = new Interactor(repository);
 		new SampleGenerator(repository).generateSamples();
-		ui = new CmdUI(repository);
+		ui = new CmdUI(interactor);
 		ui.start();
 		
 		return ostream.toString();
